wireguard client: Routing has been reworked (similar to warp), now a table is used for selective routing to remote subnets instead of ip rules, another plus is the quick removal of added rules when deleting the wg0 interface An attempt was made (not tested!) to implement ipv6 support, it may not work. The CONFIG_IPV6_MULTIPLE_TABLES parameter has been permanently added to the kernel build
nfqws: in zapret.sh the automatic detection of the wan interface has been reworked once again. Previously there could be a situation in which ipv6 interfaces with a default route with a lower priority (for example, the wg client) were used than the main one, even if the main one was already disabled. Now only those interfaces that are shown by "ip route show default" are automatically selected. If other interfaces are needed, specify them manually.